Revert "Moved builder inside system"
This reverts commit 4427f3f439f9bdd587491e5bfddeef29820f7f8f.
