+++ src/components/widgets/builder/AddWidget.svelte
