+++ src/components/widgets/builder/InputSwitch.svelte
