+++ src/components/widgets/builder/WidgetBuilder.svelte
